playwright_service/requirements.txt
aio_pika
aiofiles
alembic
asyncpg
fastapi
httpx
playwright
psycopg2-binary
pydantic-settings
python-multipart
redis
sqlalchemy
uvicorn
